Add --strict-retransmit option
So far, the session is terminated when a retransmit error is received,
as it's treated as a fatal error.  This is often too harsh, and we'd
rather want to proceed further.

So now let's be tolerant and go ahead after a retransmit error as
default.  To be more strict, pass --strict-retransmit option,
instead.