Add support of the mDNS service lookup
Add the --lookup option to the client, so that it browses mDNS and
looks for the given service name to resolve the address and the port,
instead of passing them manually.

am2n-list program shares the common code for browsing, too.

Note that the code is neither thread-safe nor optimal, it's just
provided for simplicity.