Move authentication-related code to common.c and options.c
The option handling about the authentication is moved into the
options.c while the authentication setup is moved to common.c, as code
refactoring.  The init function is simplified without checking ifdef
in the caller side.

While we're at it, make the invocation of the client easier: namely,
a client always allows all authentication methods, and it'll ask a
prompt if secret or user name isn't provided via option.  That is,
less option needed to be passed as default.