Enable the hearbeat ping to the client, too
Enable the same feature to check the liveness of the connected host with
ping for clients, too.  The same mechanism and the same options are
applied.

The only difference is the default initial timeout; it's set slightly
longer so that clients don't need to send ping unnecessarily if the host
already sends.