seq: Set the right source port at delivery
Otherwise all UMP packets are handled as if there are no source
assigned.

Ideally speaking, the kernel should take care of the right destination
port automatically when it's fed to a UMP Endpoint port, but it's not
done in that way as of this writing.  So better to set the source port
here explicitly in order to make it properly working.